Modificacion de nombre de carpetas y estructura general de alumnos y cancha
